Rough paper layout
This is a rough layout of how the paper will be set up.

1: Introduction.
Introduce the problem, etc...

2: Related work
Talk about work done by DMG to collect the information on the kernel.

3: Navigation
How our solution provides navigation mechanisms through the commit data
3.1: Search
Search mechanisms we provide.
3.2: Breadcrumbs
Breadcrumb mechanism allows users to move toward the root

4: Presentation

View files, modules, commit log, etc...

5: Date Range Queries (May be moved into navigation)

6: Tree Visualization (May be moved into navigation)
The tree provides a visual mechanism to understand location, as well as
navigate through the kernel.

...

I'm also adding a bunch of pictures of the various features in the system.