Discussion: Future Work:
We discuss the differences in our tool versus the other tool, providing a
side-by-side view of the visualization models used by our tool and gitk. We
show what information can be gathered quickly and effectively through the use
of our tool that cannot be gathered as effectively with gitk.

There is still a lot of work to be done in the future work section. We need to
talk about how our tool does not provide the diffs, though we have that
information and could very easily display it. Furthermore, we could use the
diff to determine what lines were changed, which lines were added, and which
lines were removed. We could even show what functions were edited for source
code files.